Add SupportedConfiguration schema
Gerrit:
https://gerrit.openbmc-project.xyz/c/openbmc/entity-manager/+/40293/4

Add a schema to describe supported configurations, these would
include objects that an entity supports, such as various models
of power supplies, cards, etc.
This schema would have as many optional properties as needed to
describe the supported configuration. Start with RedundantCount and
InputVoltage since the first use case would be a supported power supply.

Example, a chassis json file could have the following to describe that
it supports two power supply configurations, one is 2 2000W redundant
power supplies of the specified  model(s), the other is 4 1000W
redundant power supplies, both configurations require an input voltage
of 220V. A monitor app could then identify if any of the supported power
supplies are plugged, and then parse the entity-manager objects to find
a PowerSupply with the specified model name to compare what's supported
against what is present.

{
    "Name": "Supported Power Supply 2000W",
    "Type": "SupportedConfiguration",
    "SupportedType": "PowerSupply",
    "SupportedModel": "Delta-123",
    "RedundantCount": 2,
    "InputVoltage": [220]
},
{
    "Name": "Supported Power Supply 1000W",
    "Type": "SupportedConfiguration",
    "SupportedType": "PowerSupply",
    "SupportedModel": "Delta-890",
    "RedundantCount": 4,
    "InputVoltage": [220]
}

The above would look for an instance of
xyz.openbmc_project.Inventory.Item.PowerSupply
with a xyz.openbmc_project.Inventory.Decorator.Asset Model of any of the
listed models and get the properties for that supported configuration.
If for example a model Delta-890 is plugged, the power supply monitor
app would check that there are at least 4 present, and their input
voltage is 220V even if that power supply model can also run with 110V.

The phosphor power supply app could then have policies on how to react
based on this information. The reactions can be done via configuration
parameters in the monitoring apps. Example use cases:
- Log an error and drop system to a power safe mode and/or boost fan
  speed if number of power supplies are less than the redundant count.
- Log an error if none of the supported models are not found. This may
  indicate that the system is running in an unsupported configuration
  and the warranty may no longer apply. Also the unsupported model may
  have different power specifications (in the case of power supplies)
  that may cause issues.
- No action is taken if the system configuration matches what the EM
  supported configuration specifies.

In summary, the supported configuration schema allows the BMC to notify
the user via logging events if their system does not have the HW
configuration that is supported, and for the BMC to take actions such as
limiting the system operation because it has dropped to a safe power mode
to protect the system.

Change-Id: I630d03e61a97a2db4fe7b4e75242606f9d0e7084
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>