Add IBM PCIe cards
This adds configurations for the following cards that have similar
configs:

Flett - NVMe JBOF card
Bear River: 2U PCIe cable card
Bear Lake: 4U PCIe cable card

The TMP435s are on the same power domain as the PCIe slot itself and so
the sensor daemon needs to know that.  The PowerStateOwner JSON entry
tells hwmontempsensor this, and the LocationCode entry is used by
hwmontempsensor to find the associated PCIeSlot D-Bus object.

The TMP435s provide two temperatures, local and remote.  The remote one
is the interesting one so that is what the warning threshold is on.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: I9e1802b88389f0db13870d64378316d0505edcb4