configs:ibm: Rename PCIe card temp sensors
This commit gives the NVMe JBOF and PCIe cable card temp sensors a
stable name, so associations can be made to their corresponding D-Bus
objects in phosphor-inventory-manager (e.g.
.../motherboard/pcie_slot10/pciecard10).

The location code segment is used as the unique identifier of the sensor
names, like NVMe_JBOF_Card_C11_Temp.

The $slotnumber variable comes from the SlotNumber D-Bus property on the
inventory-manager's D-Bus object for the card.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: If98bbd5842e940fde6e3afd7f63439bc0a6dc854