EntityManager: Associate configs with their DBus probe source (#33)
Allow dynamic inventory associations to be added to sensors by reactors
in the context of our Frankenstein's monster of a PIM/EM combo.

The disconnect in the EM data model between where we know about the path of
the object that triggered the probe and where we publish the probed
configuration on DBus drives some complexity in the implementation. To
bridge that gap we need to create a bit of a wormhole through the
codebase. An alternative to making a wormhole is to inject the path
metadata straight into the EM data model, but that requires more
analysis than I have time or intestinal fortitude for right now.

The wormhole we create is built on a spiritual reimplementation of
Linux's container_of() macro. By embedding the systemConfiguration in a
struct we can reach up from its pointer and back down into another
member of the same struct which stores the probe path for a given
configuration. This is an improvement on adding another global into the
sea that we already have as globals provide no intuitive relationship
between eachother. By contrast, gathering related data into a struct
provides a much stronger intuitive relationship.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>
Change-Id: I98e137f7159113d89fe790206595955b6f311154