Add name to InventoryAdded/Removed journal entries
Add the name of the inventory item into the journal messages that print
when inventory adds or removes occur so that it's obvious which part is
being referred to.

Journal output:
entity-manager[3395]: Inventory Added: Blyth Panel

And in the journal metadata:
NAME=Blyth Panel

This also picks up the upstream change to have these be INFO traces.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: I7ce351e72831c69f3f927486350142dd466aa630