ignore unique-name dbus connections
Unique-named dbus connections should not be exposing dbus objects
and are created by dbus clients (if only an unnamed connection exists).
Filter these out in order to reduce the re-scanning effort in
entity-manager.

On some systems with a high number shell-script driven `busctl`
operations, this was reported to have a net performance benefit.  I
ran this on Bletchley hardware and observed a small (<1sec) boot time
performance, which was difficult to distinguish from the noise.

Tested: Ran on Bletchley and observed identical EM `busctl tree`.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: I08b1fe87665312e9bab019f7bff748e03cdc0027