configurations: Add Bonnell and Pennybacker
Add IBM's Bonnell chassis and its backplane Pennybacker.
The IM value for Bonnell is 50004000 as defined by the vpd repo:
```
https://github.com/openbmc/openpower-vpd-parser/blob/master/const.hpp
```
The CC value for Pennybacker is 2E44.
The sensor values for the power supplies are defined in the kernel
device tree:
```
https://github.com/openbmc/linux/blob/dev-6.0/arch/arm/boot/dts/aspeed-bmc-ibm-bonnell.dts
```

Tested: Verified that the new json files were added to the rootfs image
and that CI passed.

Change-Id: I32c8de73eb4c934da864672677a2a6dae653d241
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>