Ignore uninteresting IA/IR signals
Currently, entity-manager will parse every JSON file and do its full
D-Bus scan and probe on any interfacesAdded or interfacesRemoved signal,
without regard for which interface was added or removed.

Since the interfaces that are involved in the probes are known to
entity-manager via the Probe statements in the JSON files, it is
possible to only actually look for new or removed entities when the
signals received contain one of those interfaces.

This implements something I saw suggested in an abandoned commit[1]: to
extract the probe interfaces from the JSON files for use with the
matches.  Since the argX matches can only be done when the arg is a
string, it can't build the interfaces into the match string, but it can
check for a probe interface in the message payload before doing any real
processing.

Note that the NameOwnerChanged signal doesn't contain interfaces, only
well known bus names, so no interface checking can be done against it.

I imagine some further simplification could be done, such as not also
parsing all the JSON files as part of propertiesChangedCallback(), or at
least sharing more code, if desired.

[1] https://gerrit.openbmc.org/c/openbmc/entity-manager/+/56723

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: Iaa8d30099459d8043a49879761241cf3590c3407