Use common Polarity in schema
Point the definition of the Polarity property in the Presence entry at
the common Polarity definition, and also make that an enum since that is
how all code uses it.

Recently, CI started failing without this change, though why it started
failing is a mystery to me, and the error messages aren't that helpful.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: I966974b894a715c563bed3823a6b47b92a7ee94a