configurations: Add IBM's BlueRidge chassis
These chassis are similar enough to the Rainier ones that the Rainier
compatible names can be at the bottom of the compatible list.  This way
code that needs the name but doesn't need to tell the difference between
a Rainier and a BlueRidge doesn't have to change anything.

The IBMCompatibleSystem interface is being deprecated in favor of the
Decorator.Compatible one, but not all of the code has been changed over
yet so just use both for now.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: I453185874e43b1b5d9a215e195c74e6a439cf0d5