Add new IBM Power Mode parameters (#67)
* configurations: Add new IBM Power Mode parameters

- CustomerModes will list the power modes that are available for
customers to set on their system.
- EcoModeSupport indicates if the system will support a new set of
Efficiency based power modes.
- OemModes will list the power modes that the system supports but are
not able to be set directly by the customer. These modes may (but are
not required) to be valid Redfish PowerModes.

These parameters will be used by openpower-occ-control

Change-Id: I47e5ba17e5f0527c2cbe41cd823c21e7fedeeb58
Signed-off-by: Chris Cain <cjcain@us.ibm.com>

* configurations: Add new IBM Power Mode parms for p10 systems

- CustomerModes will list the power modes that are available for
customers to set on their system.
- EcoModeSupport indicates if the system will support a new set of
Efficiency based power modes.
- OemModes will list the power modes that the system supports but are
not able to be set directly by the customer. These modes may (but are
not required) to be valid Redfish PowerModes.

These parameters will be used by openpower-occ-control

The changes for these p10 systems is temporary and only used for early
bringup. This commit will not be upstreamed.

Change-Id: Ie80fb4b991b7501c8a657078349f872a3f37c174

---------

Signed-off-by: Chris Cain <cjcain@us.ibm.com>