Fix parent value being passed to createInterface
Confusingly, there is both a 'boardName' and 'boardNameOrig' variable
where one is "NVMe_1" and the other is "NVMe 1".  When passed into the
createInterface function, they are saved as a key to the map of D-Bus
interfaces that have been created.

When EM removes a FRU and tries to delete its interface, it does it
based on the boardNameOrig value and leaves interfaces based on
boardName, causing problems later when the FRU is added back in and the
interfaces are added back.

Fixing this in two places, though the one causing the problems was the
one used by the associations interface.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: I2641262691f8c316a5a9c65b250f6cab6ac26e07