downstream: move to lg2 and clang updates
A few miscellaneous changes to get our downstream changes working with
upstream.

The upstream clang updates changed some indents so I went with that and
then fixed the "system." aspect of our downstream calls.

Upstream moved to lg2 so needed to move our downstream logs to lg2.

Some formatting changes to our downstream commits because of the new
clang.

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>