configs: Use known num in generic_nvme_drive (#72)
Don't use $index in the D-Bus names, because EM can get confused by
platform-fru-detect interactions and tries to put two instances of the
same object path on D-Bus and will crash.

Instead use SlotNumber, which is a property on the
xyz.openbmc_project.Inventory.Decorator.Slot interface that will be on
D-Bus on the drive inventory-manager path.

Change-Id: I76b6bd5183a50a6c265bc151880c10adc7213a88

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: I178eab998e53e48c1d0823e262bcd3f21c0da136