configurations: system1: update downstream system1_chassis.json
This is the upstream version of system1_chassis.json, some of the
changes were missing downstream.

Updated the system1_baseboard.json:
Modified the PowerState for the PCIE Switches and SPYRE card
thermal sensors to only be read after the host power on.

Corrected one bus number for Spyre card.

Tested on system1 hw

Change-Id: I9fe0d0c9f8f2ba615e1f1e6dbe55cb31468f3f7f