entity-manager: Only parse config JSON once (#91)
Only parse the JSON files on the first call to loadConfigurations(). On
all future calls, just return the std::list of JSON previously read.

Previously, the files would be parsed every time
propertiesChangedCallback() was called.

Tested:
Changed the debug trace that stated the time spent during the load to
info and did a BMC reboot.  Trace only showed up once and the resulting
D-Bus tree was the same as before.


Change-Id: Id69c1a82ac550f9ef34c7a5a4ec046c65116cfa0

Signed-off-by: Matt Spinler <spinler@us.ibm.com>